FINALLY solve the node-detach-but-event-listener-still-there issue.
Multiple issues nested together: node detach from the doc but its event
listener is still on window globally, also if we use event listener
directly there is no way we can maintain "this" context as long as we
got the one single function name available in the scope for both
connectedCallback and disconnectedCallback - it's simply there's no way
we can do that.
By defining those function in connectedCallback and make it directly to
"self" instead of claiming it directly in the ES6 class, we can 1. Get
the function name globally and 2. Inside the function we can still use
"self" to indicate the element itself.